refactor: Phase-1-Verbesserungen (DI-Seams, Utils-Split, Polish, Notification-Dots)
- A3: Hilt-EntryPoint (di/LauncherServiceEntryPoint) fuer den AccessibilityService;
  AppLockManager von object zu @Singleton-Klasse (Krypto bleibt im Companion)
- A7: Icon-Cache-Build-Token als .build_token-Markerdatei statt SharedPreferences
  (inkl. einmaliger Legacy-Migration und Tests)
- A4: ui/Utils.kt zerlegt in SwipeAndBounce, SearchField, AppIconView,
  LaunchOverlays (ui/) und SystemIntegration (Root-Paket); Kover-Excludes angepasst
- C3: Weather-Refresh lifecycle-gated (repeatOnLifecycle) + pure, getestete
  Drossel-Entscheidung WeatherRepository.shouldRefresh()
- C6: StrictMode (Thread+VM) in Debug-Builds via RethroneApplication
- B3: Notification-Dots auch im App-Drawer, Toggle im Favoriten-Menue
  (ThemeManager.isNotificationDotsEnabled, Default an), Logik getestet in
  LauncherLogic.shouldShowNotificationDot; Strings in 5 Sprachen
- style: detekt --auto-correct projektweit; Baseline von 146 auf ~90 Eintraege
