+++ app/build.gradle.kts
@@ -143,8 +143,18 @@ kover {
                 classes("com.example.androidlauncher.ui.SettingsPaletteMenuKt\$*")
                 classes("com.example.androidlauncher.ui.SizeConfigMenuKt")
                 classes("com.example.androidlauncher.ui.SizeConfigMenuKt\$*")
-                classes("com.example.androidlauncher.ui.UtilsKt")
-                classes("com.example.androidlauncher.ui.UtilsKt\$*")
+                // Utils.kt wurde in vier fokussierte Dateien zerlegt (reine UI-/System-Wrapper):
+                classes("com.example.androidlauncher.ui.SwipeAndBounceKt")
+                classes("com.example.androidlauncher.ui.SwipeAndBounceKt\$*")
+                classes("com.example.androidlauncher.ui.SwipeToCloseState")
+                classes("com.example.androidlauncher.ui.SearchFieldKt")
+                classes("com.example.androidlauncher.ui.SearchFieldKt\$*")
+                classes("com.example.androidlauncher.ui.AppIconViewKt")
+                classes("com.example.androidlauncher.ui.AppIconViewKt\$*")
+                classes("com.example.androidlauncher.ui.LaunchOverlaysKt")
+                classes("com.example.androidlauncher.ui.LaunchOverlaysKt\$*")
+                classes("com.example.androidlauncher.SystemIntegrationKt")
+                classes("com.example.androidlauncher.SystemIntegrationKt\$*")
                 classes("com.example.androidlauncher.ui.ComposableSingletons*")
                 classes("com.example.androidlauncher.ComposableSingletons*")
                 classes("com.example.androidlauncher.ui.PaletteMenuItem")
