+++ app/src/main/java/com/example/androidlauncher/RethroneApplication.kt
@@ -1,6 +1,7 @@
 package com.example.androidlauncher
 
 import android.app.Application
+import android.os.StrictMode
 import dagger.hilt.android.HiltAndroidApp
 
 /**
@@ -9,4 +10,37 @@ import dagger.hilt.android.HiltAndroidApp
  * ViewModels ihre Abhaengigkeiten beziehen.
  */
 @HiltAndroidApp
-class RethroneApplication : Application()
+class RethroneApplication : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        if (BuildConfig.DEBUG) {
+            enableStrictMode()
+        }
+    }
+
+    /**
+     * StrictMode nur in Debug-Builds: macht Disk-/Netzwerk-Zugriffe auf dem Main-Thread
+     * und Ressourcen-Lecks (Activities, Closeables, Receiver) im Logcat sichtbar
+     * (Tag `StrictMode`), statt sie erst als Jank/ANR beim Nutzer zu bemerken.
+     */
+    private fun enableStrictMode() {
+        StrictMode.setThreadPolicy(
+            StrictMode.ThreadPolicy.Builder()
+                .detectDiskReads()
+                .detectDiskWrites()
+                .detectNetwork()
+                .penaltyLog()
+                .build()
+        )
+        StrictMode.setVmPolicy(
+            StrictMode.VmPolicy.Builder()
+                .detectLeakedSqlLiteObjects()
+                .detectLeakedClosableObjects()
+                .detectLeakedRegistrationObjects()
+                .detectActivityLeaks()
+                .penaltyLog()
+                .build()
+        )
+    }
+}
